Added RedisEnv & MySqlEnv
