+++ src/Option/GeneralEnv.php
@@ -6,19 +6,12 @@ namespace Nebalus\Webapi\Option;
 
 use Monolog\Level;
 
-class EnvData
+class GeneralEnv
 {
     private bool $isProduction;
     private bool $isDevelopment;
     private Level $logLevel;
     private string $jwtSecret;
-    private string $mysqlPasswd;
-    private string $mysqlHost;
-    private string $mysqlPort;
-    private string $mysqlDatabase;
-    private string $mysqlUser;
-    private string $redisHost;
-    private string $redisPort;
     private string $accessControlAllowOrigin;
     private int $jwtNormalExpirationTime;
     private int $jwtExtendedExpirationTime;
@@ -31,13 +24,6 @@ class EnvData
         $this->jwtSecret = getenv("JWT_SECRET");
         $this->jwtNormalExpirationTime = (int) getenv('JWT_NORMAL_EXPIRATION_TIME');
         $this->jwtExtendedExpirationTime = (int) getenv('JWT_EXTENDED_EXPIRATION_TIME');
-        $this->mysqlPasswd = getenv("MYSQL_PASSWORD");
-        $this->mysqlHost = getenv("MYSQL_HOST");
-        $this->mysqlPort = getenv("MYSQL_PORT");
-        $this->mysqlDatabase = getenv("MYSQL_DATABASE");
-        $this->mysqlUser = getenv("MYSQL_USER");
-        $this->redisHost = getenv("REDIS_HOST");
-        $this->redisPort = getenv("REDIS_PORT");
         $this->accessControlAllowOrigin = getenv("ACCESS_CONTROL_ALLOW_ORIGIN");
     }
 
@@ -59,42 +45,6 @@ class EnvData
     public function getJwtSecret(): string
     {
         return $this->jwtSecret;
-    }
-
-    public function getMySqlPasswd(): string
-    {
-        return $this->mysqlPasswd;
-    }
-
-    public function getMySqlHost(): string
-    {
-        return $this->mysqlHost;
-    }
-
-    public function getMySqlPort(): string
-    {
-        return $this->mysqlPort;
-    }
-
-
-    public function getMySqlDatabase(): string
-    {
-        return $this->mysqlDatabase;
-    }
-
-    public function getMySqlUser(): string
-    {
-        return $this->mysqlUser;
-    }
-
-    public function getRedisHost(): string
-    {
-        return $this->redisHost;
-    }
-
-    public function getRedisPort(): string
-    {
-        return $this->redisPort;
     }
 
     public function getJwtNormalExpirationTime(): int
